Add option to specify callback for cleanup function
The callback is called on every cached item that is up for deletion. If the callback returns true then the item is deleted. If the callback returns false then the item stays.

This is useful for, for example, writing the item to persistent storage right before the cleanup. If the write fails, for a retryable reason, then the callback can return false and it will have another shot next time the cleanup runs.